+++ engine/src/main/java/net/jqwik/engine/properties/configurators/RangeConfigurator.java
@@ -39,7 +39,7 @@ public class RangeConfigurator extends ArbitraryConfiguratorBase {
 	}
 
 	public FloatArbitrary configure(FloatArbitrary arbitrary, FloatRange range) {
-		return arbitrary.greaterOrEqual(range.min()).lessOrEqual(range.max());
+		return arbitrary.between(range.min(), range.minIncluded(), range.max(), range.maxIncluded());
 	}
 
 	public IntegerArbitrary configure(IntegerArbitrary arbitrary, IntRange range) {